+++ spimdisasm/common/ContextSymbols.py
@@ -10,10 +10,11 @@ import enum
 from typing import Callable
 import rabbitizer
 
-from .GlobalConfig import GlobalConfig, Compiler
+from .GlobalConfig import GlobalConfig
 from .FileSectionType import FileSectionType
 from .SortedDict import SortedDict
 
+from . import Utils
 
 class SymbolSpecialType(enum.Enum):
     function            = enum.auto()
@@ -246,6 +247,22 @@ class ContextSymbol:
 
     visibility: str|None = None
 
+    alignmentShift: int|None = None
+    """
+    Set an specific alignment for this symbol, possibly overriding any other
+    alignment setting from the selected compiler.
+
+    IMPORTANT: This value won't be used as-is to calculate the actual symbol
+    alignment, instead this value corresponds to "the number of low-order zero
+    bits the location counter must have after advancement", or in other words,
+    it will be used as the exponent of a power of two. For example the value `3`
+    will make this symbol to be aligned to a position multiple of `8`.
+
+    This directive will be emitted before the symbol label.
+
+    This directive won't be emitted if the symbol is not already aligned to the
+    given value.
+    """
 
     @property
     def vram(self) -> int:
@@ -724,6 +741,47 @@ class ContextSymbol:
             return f"{output} */{GlobalConfig.LINE_ENDS}"
         return ""
 
+
+    def getAlignment(self) -> int|None:
+        if self.alignmentShift is not None:
+            return 1 << self.alignmentShift
+        return None
+    def setAlignment(self, alignment: int) -> None:
+        """
+        Set an specific alignment for this symbol, possibly overriding any other
+        alignment setting from the selected compiler.
+
+        This directive will be emitted before the symbol label.
+
+        The given alignment must be a power of two (1, 2, 4, 8, 16, etc),
+        otherwise the value will be discarded.
+
+        This value will be discarded if the symbol is not already aligned to
+        the given alignment.
+        """
+        if alignment <= 0:
+            Utils.eprint(f"""
+Warning: Tried to set an alignment of `{alignment}` to symbol `{self.getName()}` (`0x{self.vram:08X}`), but the given alignment is negative.
+This alignment will be discarded.
+""")
+            return
+        # Calculate the alignment shift value
+        alignmentShift = (alignment & (-alignment)).bit_length() - 1
+        if (1 << alignmentShift) != alignment:
+            Utils.eprint(f"""
+Warning: Tried to set an alignment of `{alignment}` to symbol `{self.getName()}` (`0x{self.vram:08X}`), but the given alignment isn't a power of two.
+This alignment will be discarded.
+""")
+            return
+        alignedVram = self.vram >> alignmentShift
+        alignedVram = alignedVram << alignmentShift
+        if alignedVram != self.vram:
+            Utils.eprint(f"""
+Warning: Tried to set an alignment of `{alignment}` to symbol `{self.getName()}` (`0x{self.vram:08X}`), but the given alignment does not match the alignment the symbol currently has.
+This alignment will be discarded.
+""")
+            return
+        self.alignmentShift = alignmentShift
 
     @staticmethod
     def getCsvHeader() -> str: